--- src/components/portfolio/PfPopup.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { PfModalDiv } from "./portfolio";
-
-const PfPopup = () => {
-  return <PfModalDiv></PfModalDiv>;
-};
-
-export default PfPopup;
