+++ src/components/serviceDetail/serviceDetail.js
@@ -1,5 +1,6 @@
 import styled from "@emotion/styled";
 import { LayoutDiv } from "../../pages/page";
+import { LayerDiv, ModalDiv } from "../portfolio/portfolio";
 
 export const DetailLayout = styled(LayoutDiv)`
   display: flex;
@@ -531,6 +532,42 @@ export const ReviewDiv = styled.div`
             background-color: #eee;
           }
         }
+      }
+    }
+  }
+`;
+
+// 업체디테일 모달
+
+export const PfModalDiv = styled(ModalDiv)`
+  background-color: rgba(0, 0, 0, 0.75);
+`;
+export const PfLayerDiv = styled(LayerDiv)`
+  max-width: 680px;
+  min-height: auto;
+  height: 400px;
+  background-color: #fff;
+  display: flex;
+  padding: 50px 0;
+
+  .photo-area {
+    border: 1px solid;
+    width: 400px;
+    height: 100%;
+  }
+
+  .txt-area {
+    border: 1px solid;
+    width: calc(100% - 400px);
+    height: 100%;
+
+    h3 {
+      font-size: 24px;
+      font-weight: 600;
+    }
+    ul {
+      border: 1px solid #eee;
+      li {
       }
     }
   }
